sentinel_x_v01/requirements.txt
alpaca-py>=0.14.0
fastapi>=0.100.0
uvicorn>=0.22.0
